move MapFeedbackChip to chips module
